ai-agent/pyproject.toml
[project]
name = "visulate-ai-agent"
version = "2.1.0"
description = "AI Agent for Oracle database interactions using Visulate"
readme = "README.md"
requires-python = ">=3.10"
authors = [
    {name = "Visulate LLC", email = "support@visulate.com"},
]
keywords = ["oracle", "database", "ai", "agent", "mcp", "google-adk"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Database",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
    "a2a-sdk==0.3.3",
    "fastmcp==2.14.0",
    "google-adk==1.13.0",
    "google-genai>=1.17.0",
    "python-dotenv>=1.1.0",
    "requests>=2.31.0",
    "uvicorn>=0.20.0",
    "fastapi>=0.100.0",
    "pytest>=7.0.0",
    "pytest-asyncio>=0.20.0",
    "httpx>=0.24.0",
]

[project.scripts]
visulate-agent = "root_agent.main:main"

[tool.setuptools]
packages = ["common", "root_agent", "nl2sql_agent", "comment_generator", "object_analysis_agent", "erd_agent", "schema_analysis_agent"]

[project.urls]
Homepage = "https://visulate.net"
Documentation = "https://github.com/visulate/visulate-for-oracle/blob/main/ai-agent/README.md"
Repository = "https://github.com/visulate/visulate-for-oracle.git"